Time: 1 ms (100%), Space: 62.1 MB (65.44%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.java
@@ -9,6 +9,6 @@ class Solution {
             else if(nums[i] > max) max = nums[i];
         }
 
-        return k * (max - min);
+        return (long) k * (max - min);
     }
 }
